Move all visualization views, URLs, and templates to visualizations app
- Move VisualizationListView and VisualizationDetailView from pages
- Move census map views (denomination, demographics, geojson, urban
  congregations) from census/views.py
- Move datalayer map views from datalayers/views.py
- Consolidate all /visualizations/ URL routes in visualizations/urls.py
- Simplify list template: use get_absolute_url() instead of slug checks
- Remove DataLayerSourceAdmin (consolidated into VisualizationAdmin)
- Update all Visualization imports in sitemaps and management commands
- Fix duplicate pages.urls include in config/urls.py
- Add VisualizationAdmin with Unfold styling and render_type fields
