+++ pages/views.py
@@ -1,14 +1,11 @@
-import json
-
 import markdown
-from django.conf import settings
 from django.http import Http404
 from django.shortcuts import get_object_or_404, render
 from django.utils.decorators import method_decorator
 from django.views.decorators.cache import cache_page
 from django.views.generic import DetailView, ListView
 
-from .models import BlogPost, Page, Visualization
+from .models import BlogPost, Page
 
 
 # Static Pages
@@ -22,7 +19,6 @@ def page_detail(request, slug):
         if not (request.user.is_authenticated and request.user.is_staff):
             raise Http404("Page not found")
 
-    # Set page title for template
     context = {
         "page": page,
         "page_title": page.title,
@@ -59,7 +55,7 @@ def nav_pages_context(request):
 
 
 # Blog Posts
-@method_decorator(cache_page(60 * 15), name="dispatch")  # 15 minutes
+@method_decorator(cache_page(60 * 15), name="dispatch")
 class BlogListView(ListView):
     model = BlogPost
     template_name = "pages/blog_list.html"
@@ -70,7 +66,7 @@ class BlogListView(ListView):
         return BlogPost.objects.filter(is_draft=False).order_by("-published_date")
 
 
-@method_decorator(cache_page(60 * 15), name="dispatch")  # 15 minutes
+@method_decorator(cache_page(60 * 15), name="dispatch")
 class BlogDetailView(DetailView):
     model = BlogPost
     template_name = "pages/blog_detail.html"
@@ -82,109 +78,8 @@ class BlogDetailView(DetailView):
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
-        # Render markdown to HTML
         md = markdown.Markdown(
-            extensions=[
-                "extra",  # Tables, footnotes, etc.
-                "codehilite",  # Code highlighting
-                "toc",  # Table of contents
-            ]
+            extensions=["extra", "codehilite", "toc"]
         )
         context["content_html"] = md.convert(self.object.content)
-        return context
-
-
-# Visualizations
-@method_decorator(cache_page(60 * 15), name="dispatch")  # 15 minutes
-class VisualizationListView(ListView):
-    model = Visualization
-    template_name = "pages/visualization_list.html"
-    context_object_name = "visualizations"
-    paginate_by = 10
-
-    def get_queryset(self):
-        return Visualization.objects.all().order_by("-published_date")
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        from django.db.models import Count
-
-        from datalayers.models import DataLayer, DataLayerSource
-
-        # Get data layer sources with point counts
-        datalayer_qs = (
-            DataLayer.objects.filter(lat__isnull=False, lon__isnull=False)
-            .values("source")
-            .annotate(count=Count("id"))
-            .order_by("source")
-        )
-        # Look up source metadata for titles, thumbnails, etc.
-        source_meta = {s.slug: s for s in DataLayerSource.objects.all()}
-        context["datalayer_sources"] = [
-            {
-                **dl,
-                "meta": source_meta.get(dl["source"]),
-                "display_title": (
-                    source_meta[dl["source"]].title
-                    if dl["source"] in source_meta
-                    else dl["source"].replace("-", " ").title()
-                ),
-            }
-            for dl in datalayer_qs
-        ]
-        return context
-
-
-@method_decorator(cache_page(60 * 15), name="dispatch")  # 15 minutes
-class VisualizationDetailView(DetailView):
-    model = Visualization
-    template_name = "pages/visualization_detail.html"
-    context_object_name = "visualization"
-    slug_field = "slug"
-
-    def get_template_names(self):
-        """Check for custom template based on slug, fall back to default"""
-        custom_template = f"pages/visualizations/{self.object.slug}.html"
-        # Try custom template first, then fall back to default
-        return [custom_template, self.template_name]
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        # Render markdown to HTML
-        md = markdown.Markdown(
-            extensions=[
-                "extra",
-                "codehilite",
-                "toc",
-            ]
-        )
-        context["content_html"] = md.convert(self.object.content)
-
-        # Load JSON data for specific visualizations
-        if self.object.slug == "catholic-dioceses":
-            # Try STATICFILES_DIRS first (development), then STATIC_ROOT (production)
-            if hasattr(settings, "STATICFILES_DIRS") and settings.STATICFILES_DIRS:
-                base_path = settings.STATICFILES_DIRS[0]
-            else:
-                base_path = settings.STATIC_ROOT
-
-            dioceses_path = base_path / "data" / "catholic_dioceses.json"
-
-            try:
-                with open(dioceses_path, "r", encoding="utf-8") as f:
-                    # Load and re-serialize to JSON string for template
-                    context["dioceses_data"] = json.dumps(json.load(f))
-
-                # Use CHNM API for historical state boundaries instead of embedding large GeoJSON
-                # The template will fetch from: http://data.chnm.org/ahcb/states/[date]/
-                context["use_chnm_api_for_map"] = True
-            except (FileNotFoundError, json.JSONDecodeError) as e:
-                # Gracefully handle missing or invalid data files
-                context["dioceses_data"] = "[]"
-                context["use_chnm_api_for_map"] = False
-                # Log the error in development
-                import logging
-
-                logging.warning(f"Error loading Catholic dioceses data: {e}")
-
         return context
