+++ templates/visualizations/visualization_detail.html
@@ -1,0 +1,178 @@
+{% extends "base.html" %}
+{% load static %}
+
+{% block title %}{{ visualization.title }} - Religious Ecologies{% endblock %}
+
+{% block extra_css %}
+<!-- Visualization-specific CSS -->
+    {% if visualization.style_file %}
+        <link rel="stylesheet" href="{% static visualization.style_file %}">
+    {% endif %}
+    <!-- Leaflet CSS for map visualizations -->
+    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css"
+          integrity="sha256-p4NxAoJBhIIN+hmNHrzRCf9tD/miZyoHS5obTRR9BMY="
+          crossorigin="" />
+    <style>
+    /* Markdown content styling */
+        .markdown-content {
+            line-height: 1.8;
+        }
+        .markdown-content h2 {
+            font-size: 1.875rem;
+            font-weight: 700;
+            margin-top: 2rem;
+            margin-bottom: 1rem;
+            color: #1f2937;
+        }
+        .markdown-content h3 {
+            font-size: 1.5rem;
+            font-weight: 600;
+            margin-top: 1.5rem;
+            margin-bottom: 0.75rem;
+            color: #374151;
+        }
+        .markdown-content p {
+            margin-bottom: 1rem;
+            color: #374151;
+        }
+        .markdown-content ul, .markdown-content ol {
+            margin-bottom: 1rem;
+            padding-left: 2rem;
+        }
+        .markdown-content li {
+            margin-bottom: 0.5rem;
+        }
+        .markdown-content a {
+            color: #4f46e5;
+            text-decoration: underline;
+        }
+        .markdown-content a:hover {
+            color: #4338ca;
+        }
+        .markdown-content blockquote {
+            border-left: 4px solid #e5e7eb;
+            padding-left: 1rem;
+            margin: 1.5rem 0;
+            color: #6b7280;
+            font-style: italic;
+        }
+        .markdown-content code {
+            background-color: #f3f4f6;
+            padding: 0.2rem 0.4rem;
+            border-radius: 0.25rem;
+            font-family: monospace;
+            font-size: 0.875em;
+        }
+        .markdown-content pre {
+            background-color: #1f2937;
+            color: #f9fafb;
+            padding: 1rem;
+            border-radius: 0.5rem;
+            overflow-x: auto;
+            margin: 1rem 0;
+        }
+        .markdown-content pre code {
+            background-color: transparent;
+            padding: 0;
+            color: inherit;
+        }
+        .markdown-content table {
+            width: 100%;
+            border-collapse: collapse;
+            margin: 1.5rem 0;
+        }
+        .markdown-content th, .markdown-content td {
+            border: 1px solid #e5e7eb;
+            padding: 0.75rem;
+            text-align: left;
+        }
+        .markdown-content th {
+            background-color: #f9fafb;
+            font-weight: 600;
+        }
+    /* Viz-specific container styling */
+        .viz-block {
+            margin: 2rem 0;
+        }
+        #chrono-map {
+            width: 100%;
+            height: 600px;
+        }
+    </style>
+{% endblock %}
+
+{% block content %}
+    <article class="max-w-7xl mx-auto">
+    <!-- Header -->
+        <header class="mb-8">
+            <h1 class="text-4xl font-bold text-gray-900 mb-4">{{ visualization.title }}</h1>
+            <div class="flex flex-wrap items-center text-gray-600 text-sm mb-4">
+                {% if visualization.author %}
+                    <span class="font-medium">{{ visualization.author }}</span>
+                    <span class="mx-3">•</span>
+                {% endif %}
+                <time datetime="{{ visualization.published_date|date:'Y-m-d' }}">
+                    Published: {{ visualization.published_date|date:"F j, Y" }}
+                </time>
+                {% if visualization.updated_date and visualization.updated_date != visualization.published_date %}
+                    <span class="mx-3">•</span>
+                    <time datetime="{{ visualization.updated_date|date:'Y-m-d' }}">
+                        Updated: {{ visualization.updated_date|date:"F j, Y" }}
+                    </time>
+                {% endif %}
+                {% if visualization.doi %}
+                    <span class="mx-3">•</span>
+                    <a href="{{ visualization.doi }}"
+                       target="_blank"
+                       rel="noopener noreferrer"
+                       class="text-indigo-600 hover:text-indigo-700">
+                        DOI: {{ visualization.doi|slice:"18:" }}
+                    </a>
+                {% endif %}
+            </div>
+            {% if visualization.abstract %}
+                <div class="p-4 mb-6">
+                    <p class="text-gray-700 text-2xl">{{ visualization.abstract }}</p>
+                </div>
+            {% endif %}
+        </header>
+
+    <!-- Content (including embedded visualization HTML) -->
+        <div class="markdown-content prose prose-lg max-w-none mb-8">
+            {{ content_html|safe }}
+        </div>
+
+    <!-- Footer -->
+        <footer class="mt-12 pt-8 border-t border-gray-200">
+            <a href="{% url 'visualization-list' %}"
+               class="inline-flex items-center text-indigo-600 hover:text-indigo-700 font-medium">
+                <svg class="w-5 h-5 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+                    <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10 19l-7-7m0 0l7-7m-7 7h18"></path>
+                </svg>
+                Back to Visualizations
+            </a>
+        </footer>
+    </article>
+{% endblock %}
+
+{% block extra_js %}
+<!-- Import maps to resolve module imports from CDN -->
+    <script type="importmap">
+        {
+            "imports": {
+                "d3": "https://cdn.jsdelivr.net/npm/d3@7/+esm",
+                "@observablehq/plot": "https://cdn.jsdelivr.net/npm/@observablehq/plot@0.6/+esm",
+                "leaflet": "https://cdn.jsdelivr.net/npm/leaflet@1.9.4/+esm",
+                "@params": "{% static 'viz/params.js' %}"
+            }
+        }
+    </script>
+<!-- Leaflet JS for map visualizations -->
+    <script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js"
+            integrity="sha256-20nQCchB9co0qIjJZRGuk2/Z9VM+kNiyxNV1lvTlZBo="
+            crossorigin=""></script>
+<!-- Visualization-specific JavaScript -->
+    {% if visualization.script_file %}
+        <script type="module" src="{% static visualization.script_file %}?v={{ visualization.updated_at|date:'U' }}"></script>
+    {% endif %}
+{% endblock %}
