+++ templates/visualizations/visualization_list.html
@@ -1,0 +1,107 @@
+{% extends "base.html" %}
+{% load static %}
+
+{% block title %}Visualizations - Religious Ecologies{% endblock %}
+
+{% block content %}
+    <div class="max-w-6xl mx-auto">
+        <h1 class="text-4xl font-bold text-gray-900 mb-8">Visualizations</h1>
+
+        {% if visualizations %}
+            <div class="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-8">
+                {% for viz in visualizations %}
+                    <article class="bg-white shadow-sm rounded-lg overflow-hidden hover:shadow-md transition duration-150">
+                        <a href="{{ viz.get_absolute_url }}" class="block">
+                            {% with thumbnail_url=viz.get_thumbnail_url %}
+                                {% if thumbnail_url %}
+                                    <img src="{{ thumbnail_url }}"
+                                         alt="{{ viz.thumbnail_description|default:viz.title }}"
+                                         class="w-full h-48 object-cover">
+                                {% else %}
+                                    <div class="w-full h-48 bg-gray-100 flex items-center justify-center">
+                                        <svg class="w-16 h-16 text-gray-300" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                                            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="1.5"
+                                                  d="M17.657 16.657L13.414 20.9a1.998 1.998 0 01-2.827 0l-4.244-4.243a8 8 0 1111.314 0z"/>
+                                            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="1.5"
+                                                  d="M15 11a3 3 0 11-6 0 3 3 0 016 0z"/>
+                                        </svg>
+                                    </div>
+                                {% endif %}
+                            {% endwith %}
+                            <div class="p-6">
+                                <h2 class="text-xl font-bold text-gray-900 mb-2 hover:text-indigo-600 transition">
+                                    {{ viz.title }}
+                                </h2>
+                                <div class="text-sm text-gray-600 mb-3">
+                                    {% if viz.author %}
+                                        <span class="font-medium">{{ viz.author }}</span>
+                                        <span class="mx-2">&bull;</span>
+                                    {% endif %}
+                                    <time datetime="{{ viz.published_date|date:'Y-m-d' }}">
+                                        {{ viz.published_date|date:"F j, Y" }}
+                                    </time>
+                                    {% if viz.updated_date and viz.updated_date != viz.published_date %}
+                                        <span class="mx-2">&bull;</span>
+                                        <span>Updated {{ viz.updated_date|date:"F j, Y" }}</span>
+                                    {% endif %}
+                                </div>
+                                {% if viz.abstract %}
+                                    <p class="text-gray-700 text-sm leading-relaxed">
+                                        {{ viz.abstract|truncatewords:30 }}
+                                    </p>
+                                {% endif %}
+                                {% if viz.doi %}
+                                    <div class="mt-3 text-xs text-gray-500">
+                                        DOI: {{ viz.doi|slice:"18:" }}
+                                    </div>
+                                {% endif %}
+                            </div>
+                        </a>
+                    </article>
+                {% endfor %}
+            </div>
+
+            {% if is_paginated %}
+                <nav class="mt-8 flex justify-center">
+                    <ul class="flex space-x-2">
+                        {% if page_obj.has_previous %}
+                            <li>
+                                <a href="?page=1"
+                                   class="px-3 py-2 bg-white border border-gray-300 rounded-md hover:bg-gray-50">
+                                    First
+                                </a>
+                            </li>
+                            <li>
+                                <a href="?page={{ page_obj.previous_page_number }}"
+                                   class="px-3 py-2 bg-white border border-gray-300 rounded-md hover:bg-gray-50">
+                                    Previous
+                                </a>
+                            </li>
+                        {% endif %}
+
+                        <li class="px-3 py-2 bg-indigo-600 text-white rounded-md">
+                            Page {{ page_obj.number }} of {{ page_obj.paginator.num_pages }}
+                        </li>
+
+                        {% if page_obj.has_next %}
+                            <li>
+                                <a href="?page={{ page_obj.next_page_number }}"
+                                   class="px-3 py-2 bg-white border border-gray-300 rounded-md hover:bg-gray-50">
+                                    Next
+                                </a>
+                            </li>
+                            <li>
+                                <a href="?page={{ page_obj.paginator.num_pages }}"
+                                   class="px-3 py-2 bg-white border border-gray-300 rounded-md hover:bg-gray-50">
+                                    Last
+                                </a>
+                            </li>
+                        {% endif %}
+                    </ul>
+                </nav>
+            {% endif %}
+        {% else %}
+            <p class="text-gray-600">No visualizations available.</p>
+        {% endif %}
+    </div>
+{% endblock %}
